Move chunkArray to a util
