+++ app/utils/chunk-array.ts
@@ -1,0 +1,10 @@
+export default function chunkArray(arr: any[], limit: number): any[][] {
+    const original = arr.slice();
+    const result: any[][] = [];
+
+    while (original.length) {
+        result.push(original.splice(0, limit));
+    }
+
+    return result;
+}
