+++ app/validators/password-strength.ts
@@ -16,22 +16,4 @@ const PasswordStrength = BaseValidator.extend({
     },
 });
 
-// PasswordStrength.reopenClass({
-//     /**
-//      * Define attribute specific dependent keys for your validator
-//      *
-//      * [
-//      *     `model.array.@each.${attribute}` --> Dependent is created on the model's context
-//      *     `${attribute}.isValid` --> Dependent is created on the `model.validations.attrs` context
-//      * ]
-//      *
-//      * @param {String} attribute The attribute being evaluated
-//      * @param {Unknown} options  Options passed into your validator
-//      * @return {Array}
-//      */
-//     getDependentsFor(attribute: string, options: object): any[] {
-//         return [];
-//     },
-// });
-
 export default PasswordStrength;
